requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-swagger==1.0.16
aiosignal==1.3.1
anyio==3.6.2
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dill==0.3.6
distro==1.8.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
frozenlist==1.3.3
future==0.18.3
genie==23.4
genie.libs.clean==23.4
genie.libs.conf==23.4
genie.libs.filetransferutils==23.4
genie.libs.health==23.4
genie.libs.ops==23.4
genie.libs.parser==23.4
genie.libs.sdk==23.4
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
junit-xml==1.9
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
ncclient==0.6.13
netaddr==0.8.0
netmiko==4.2.0
ntc-templates==3.3.0
orjson==3.8.12
packaging==23.1
paramiko==3.1.0
pathspec==0.11.1
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.5
pyats==23.4
pyats.aereport==23.4
pyats.aetest==23.4
pyats.async==23.4
pyats.connections==23.4
pyats.datastructures==23.4
pyats.easypy==23.4
pyats.kleenex==23.4
pyats.log==23.4
pyats.reporter==23.4
pyats.results==23.4
pyats.tcl==23.4
pyats.topology==23.4
pyats.utils==23.4
pycparser==2.21
pydantic==1.10.7
pyftpdlib==1.5.7
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.0
python-engineio==3.14.2
python-multipart==0.0.6
python-socketio==4.6.1
PyYAML==6.0
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scp==0.14.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
textfsm==1.1.3
tftpy==0.8.0
tqdm==4.65.0
typing_extensions==4.5.0
ujson==5.7.0
unicon==23.4
unicon.plugins==23.4
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
xmltodict==0.13.0
yamllint==1.31.0
yang.connector==23.4
yarl==1.9.2
